Adds conditions to installer to skip sections which have already been run

+++ install.sh
@@ -19,33 +19,42 @@ apt-get update
 apt-get install -y vlc ffmpeg php8.2-cli php8.2-sqlite php8.2-curl php8.2-xml git unzip
 
 # Install Yt-dlp
-if [[ "$(dpkg --print-architecture)" == "amd64" ]]; then
-    curl -L https://github.com/yt-dlp/yt-dlp/releases/latest/download/yt-dlp_linux -o /usr/local/bin/yt-dlp
-elif [[ "$(dpkg --print-architecture)" == "armhf" ]]; then
-    curl -L https://github.com/yt-dlp/yt-dlp/releases/latest/download/yt-dlp_linux_armv7l -o /usr/local/bin/yt-dlp
-else
-    echo "Error: unknown arch."
-    echo "Please install Yt-dlp manually: https://github.com/yt-dlp/yt-dlp/releases/latest/"
-    exit 1
+if [[ ! -f /usr/local/bin/yt-dlp ]]; then
+    if [[ "$(dpkg --print-architecture)" == "amd64" ]]; then
+        curl -L https://github.com/yt-dlp/yt-dlp/releases/latest/download/yt-dlp_linux -o /usr/local/bin/yt-dlp
+    elif [[ "$(dpkg --print-architecture)" == "armhf" ]]; then
+        curl -L https://github.com/yt-dlp/yt-dlp/releases/latest/download/yt-dlp_linux_armv7l -o /usr/local/bin/yt-dlp
+    else
+        echo "Error: unknown arch."
+        echo "Please install Yt-dlp manually: https://github.com/yt-dlp/yt-dlp/releases/latest/"
+        exit 1
+    fi
+
+    chmod a+rx /usr/local/bin/yt-dlp
 fi
 
-chmod a+rx /usr/local/bin/yt-dlp
 
 # Install Composer
 curl -sS https://getcomposer.org/installer | /usr/bin/php -- --install-dir=/usr/bin --filename=composer
 chmod 0755 /usr/bin/composer
 
 # Install dependencies
-composer install
-#php artisan octane:install
+if [[ ! -d "${__DIR__}/vendor" ]]; then
+    composer install
+    #php artisan octane:install
+fi
 
 # Configure project
-cp .env.example .env
-php artisan key:generate
+if [[ ! -f "${__DIR__}/.env" ]]; then
+    cp .env.example .env
+    php artisan key:generate
+fi
 
 # Initiate database
-touch database/database.sqlite
-php artisan migrate
+if [[ ! -f "${__DIR__}/database/database.sqlite" ]]; then
+    touch database/database.sqlite
+    php artisan migrate
+fi
 
 # Install services
 sed "s|{{ PWD }}|${__DIR__}|g" "${__DIR__}/stubs/systemctl/jukebox-queue-worker.service" > /etc/systemd/system/jukebox-queue-worker.service
